// src/contracts/types/web3-contracts/MigratedZBGCard.d.ts
export class MigratedZBGCard extends Contract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  );
  methods: {
    supportsInterface(
      _interfaceId: string | number[]
    ): TransactionObject<boolean>;

    getApproved(_tokenId: number | string): TransactionObject<string>;

    individualSupply(_tokenId: number | string): TransactionObject<string>;

    tokensOwned(
      _owner: string
    ): TransactionObject<{
      indexes: (string)[];
      balances: (string)[];
      0: (string)[];
      1: (string)[];
    }>;
// ...
  };
// ...
}